Senior_Practical Cloud Computing Platform/Final Project/src_Azure-Function/Back-End/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# python -m pip install --upgrade pip
# python -m pip install -r requirements.txt
azure-functions
azure-storage-queue
azure-cosmos